Move cdm module under io.
