+++ metpy/io/cdm.py
